Move part of Slice to Core/Slice
