+++ src/Fragnix/Slice.hs
@@ -25,13 +25,9 @@ module Fragnix.Slice
 
 import Prelude hiding (writeFile,readFile)
 
-import Data.Aeson (FromJSON, ToJSON, eitherDecode)
+import Data.Aeson (eitherDecode)
 import Data.Aeson.Encode.Pretty (encodePretty)
 
-import GHC.Generics (Generic)
-import Data.Hashable (Hashable)
-
-import Data.Text (Text)
 import qualified Data.Text as Text (unpack,pack,length,index)
 
 import Control.Monad.Trans.State.Strict (StateT,execStateT,get,put)
@@ -44,150 +40,8 @@ import Data.ByteString.Lazy (writeFile,readFile)
 import System.FilePath ((</>),dropFileName)
 import System.Directory (createDirectoryIfMissing)
 import Data.Char (isDigit)
+import Fragnix.Core.Slice
 import Fragnix.Utils (listFilesRecursive)
-
--- Slices
-
-data Slice = Slice
-  { sliceID :: SliceID
-  , language :: Language
-  , fragment :: Fragment
-  , uses :: [Use]
-  , instances :: [Instance]
-  }
-deriving instance Show Slice
-deriving instance Eq Slice
-deriving instance Ord Slice
-deriving instance Generic Slice
-instance ToJSON Slice
-instance FromJSON Slice
-
--- Language
-
-data Language = Language { extensions :: [GHCExtension] }
-
-deriving instance Show Language
-deriving instance Eq Language
-deriving instance Ord Language
-deriving instance Generic Language
-instance ToJSON Language
-instance FromJSON Language
-instance Hashable Language
-
--- Fragment
-
-data Fragment = Fragment [SourceCode]
-
-deriving instance Show Fragment
-deriving instance Eq Fragment
-deriving instance Ord Fragment
-deriving instance Generic Fragment
-
-instance ToJSON Fragment
-instance FromJSON Fragment
-instance Hashable Fragment
-
--- Use
-
-data Use = Use
-  { qualification :: (Maybe Qualification)
-  , usedName :: UsedName
-  , reference :: Reference
-  }
-
-deriving instance Show Use
-deriving instance Eq Use
-deriving instance Ord Use
-deriving instance Generic Use
-
-instance ToJSON Use
-instance FromJSON Use
-instance Hashable Use
-
--- Instance
-
-data Instance = Instance
-  { instancePart :: InstancePart
-  , instanceID :: InstanceID
-  }
-
-deriving instance Show Instance
-deriving instance Eq Instance
-deriving instance Ord Instance
-deriving instance Generic Instance
-
-instance ToJSON Instance
-instance FromJSON Instance
-instance Hashable Instance
-
--- Instance Part
-
-data InstancePart =
-    OfThisClass |
-    OfThisClassForUnknownType |
-    ForThisType |
-    ForThisTypeOfUnknownClass
-
-deriving instance Show InstancePart
-deriving instance Eq InstancePart
-deriving instance Ord InstancePart
-deriving instance Generic InstancePart
-
-instance ToJSON InstancePart
-instance FromJSON InstancePart
-instance Hashable InstancePart
-
--- UsedName
-
-data UsedName =
-    ValueName { valueName :: Name } |
-    TypeName { typeName :: Name } |
-    ConstructorName { constructorTypeName :: TypeName, constructorName :: Name }
-
-deriving instance Show UsedName
-deriving instance Eq UsedName
-deriving instance Ord UsedName
-deriving instance Generic UsedName
-
-instance ToJSON UsedName
-instance FromJSON UsedName
-instance Hashable UsedName
-
--- Reference
-
-data Reference = OtherSlice SliceID | Builtin OriginalModule
-
-deriving instance Show Reference
-deriving instance Eq Reference
-deriving instance Ord Reference
-deriving instance Generic Reference
-
-instance ToJSON Reference
-instance FromJSON Reference
-instance Hashable Reference
-
--- Name
-
-data Name = Identifier Text | Operator Text
-
--- Name instances
-
-deriving instance Show Name
-deriving instance Eq Name
-deriving instance Ord Name
-deriving instance Generic Name
-
-instance ToJSON Name
-instance FromJSON Name
-instance Hashable Name
-
-type InstanceID = SliceID
-type TypeName = Name
-type SliceID = Text
-type SourceCode = Text
-type Qualification = Text
-type OriginalModule = Text
-type GHCExtension = Text
 
 -- Slice parse errors
